book integration
