+++ src/Assets/homepage.scss
@@ -47,7 +47,7 @@ $boxFont: 'Trebuchet MS', 'Lucida Sans Unicode', 'Lucida Grande', 'Lucida Sans',
   }
   .nav-links{
     margin-top: 10px;
-    .log_out{
+    .log_out {
       position: absolute;
       bottom: 0;
       width: $defaultWidth;
