+++ src/Assets/profile.scss
@@ -164,7 +164,7 @@
               font-weight: 600;
               color: grey;
             }
-            p2 {
+            p {
               font-family: verdana;
               letter-spacing: 1px;
               font-weight: 300;
