+++ src/utils/useSignUp.js
@@ -4,32 +4,24 @@ import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Api } from '../app';
 
-const useSignIn = () => {
+const useSignUp = () => {
   const redirect = useNavigate();
   const [form, setForm] = useState({});
   const [error, setError] = useState('');
-  const [passwordicon, setPasswordIcon] = useState(false);
 
-  const formInput = (e) => {
+  // onChange function
+  const handleFormValue = (e) => {
     setForm({ ...form, [e.target.name]: e.target.value });
   };
 
-  // login field
-  const switchField = (input) => {
-    if (input.current.type === 'password') {
-      input.current.type = 'text';
-      setPasswordIcon(true);
-    } else {
-      input.current.type = 'password';
-      setPasswordIcon(false);
-    }
-  };
-
-  const logInUser = async () => {
+  // Register function
+  const registerUser = async () => {
     try {
-      const res = await Api.post('/signin', {
+      const res = await Api.post('/signup', {
+        userName: form.username,
         email: form.email,
         password: form.password,
+        reEnterPassword: form.confirmPassword,
       });
       sessionStorage.setItem('user', JSON.stringify(res.data));
       setError('');
@@ -39,13 +31,11 @@ const useSignIn = () => {
     }
   };
 
-  return [
+  return {
     error,
-    passwordicon,
-    switchField,
-    formInput,
-    logInUser,
-  ];
+    handleFormValue,
+    registerUser,
+  };
 };
 
-export default useSignIn;
+export default useSignUp;
